Implemented the music and buttons to the game activity.
