+++ app/src/main/AndroidManifest.xml
@@ -15,7 +15,6 @@
 
             </intent-filter>
         </activity>
-        <activity android:name=".MainActivity" />
         <activity android:name=".MainMenuActivity" />
         <activity android:name=".HighScoresActivity"></activity>
         <activity android:name=".Game4x4Activity"></activity>
