+++ app/src/main/java/com/example/samlee/cs245application/Game4x4Activity.java
@@ -1,13 +1,15 @@
 package com.example.samlee.cs245application;
 
+import android.content.Intent;
+import android.media.MediaPlayer;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 import android.view.View;
+import android.widget.CheckBox;
+import android.widget.CompoundButton;
 import android.widget.GridLayout;
 import android.os.Handler;
-
 import java.util.Random;
-import java.util.logging.LogRecord;
 
 public class Game4x4Activity extends AppCompatActivity implements View.OnClickListener{
 
@@ -22,6 +24,9 @@ public class Game4x4Activity extends AppCompatActivity implements View.OnClickLi
     private MemoryButton selectedButton2;
 
     private boolean isBusy = false;
+
+    private int paused;
+    private MediaPlayer player;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -69,6 +74,24 @@ public class Game4x4Activity extends AppCompatActivity implements View.OnClickLi
             }
             j++;
         }
+
+        player = MediaPlayer.create(this, R.raw.get_schwifty);
+        player.setLooping(true);
+        player.start();
+
+        CheckBox checkbox = (CheckBox) findViewById(R.id.pauseCheckBox);
+        checkbox.setOnCheckedChangeListener(new CompoundButton.OnCheckedChangeListener() {
+            @Override
+            public void onCheckedChanged(CompoundButton compoundButton, boolean isChecked) {
+                if(compoundButton.isChecked()) {
+                    paused = player.getCurrentPosition();
+                    player.pause();
+                } else {
+                    player.seekTo(paused);
+                    player.start();
+                }
+            }
+        });
     }
 
     protected void shuffleButtonGraphics()
@@ -149,5 +172,26 @@ public class Game4x4Activity extends AppCompatActivity implements View.OnClickLi
                 }
             }, 500);
         }
+    }
+
+    public void tryAgainClicked(View view){
+
+    }
+
+    public void newGameClicked(View view){
+
+    }
+
+    public void endGameClicked(View view){
+        startActivity(new Intent(this, MainMenuActivity.class));
+        player.release();
+        finish();
+    }
+
+    @Override
+    protected void onPause() {
+        super.onPause();
+        player.release();
+        finish();
     }
 }
