+++ app/src/main/res/layout/activity_game4x4.xml
@@ -1,15 +1,77 @@
 <?xml version="1.0" encoding="utf-8"?>
-<GridLayout xmlns:android="http://schemas.android.com/apk/res/android"
+
+<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
+    android:id="@+id/activity_main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:paddingBottom="@dimen/activity_vertical_margin"
     android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context="com.example.samlee.cs245application.Game4x4Activity"
-    android:id="@+id/grid_layout_4x4"
-    android:columnCount="4"
-    android:rowCount="5">
+    tools:context="com.example.samlee.cs245application.Game4x4Activity">
 
-</GridLayout>
+    <Button
+        android:text="@string/try_again"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:onClick="tryAgainClicked"
+        android:id="@+id/tryAgain"
+        tools:ignore="RtlHardcoded"
+        android:layout_alignParentTop="true"
+        android:layout_alignStart="@+id/newGame" />
+
+    <Button
+        android:text="@string/end_game"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_alignParentBottom="true"
+        android:layout_alignParentRight="true"
+        android:onClick="endGameClicked"
+        android:id="@+id/endGame"
+        android:layout_alignParentEnd="true"
+        tools:ignore="RtlHardcoded" />
+
+    <Button
+        android:text="@string/new_game"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:onClick="newGameClicked"
+        android:id="@+id/newGame"
+        android:layout_marginBottom="59dp"
+        android:layout_above="@+id/endGame"
+        android:layout_alignParentRight="true"
+        android:layout_alignParentEnd="true"
+        tools:ignore="RtlHardcoded" />
+
+    <CheckBox
+        android:text="@string/mute_game_music"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_marginTop="68dp"
+        android:id="@+id/pauseCheckBox"
+        android:textAlignment="viewStart"
+        android:textSize="12sp"
+        android:layout_below="@+id/tryAgain"
+        android:layout_alignParentRight="true"
+        android:layout_alignParentEnd="true"
+        tools:ignore="RtlHardcoded" />
+
+    <GridLayout xmlns:android="http://schemas.android.com/apk/res/android"
+        xmlns:tools="http://schemas.android.com/tools"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        android:paddingBottom="@dimen/activity_vertical_margin"
+        android:paddingLeft="@dimen/activity_horizontal_margin"
+        android:paddingRight="@dimen/activity_horizontal_margin"
+        android:paddingTop="@dimen/activity_vertical_margin"
+        tools:context="com.example.samlee.cs245application.Game4x4Activity"
+        android:id="@+id/grid_layout_4x4"
+        android:columnCount="4"
+        android:rowCount="5"
+        android:layout_toStartOf="@+id/pauseCheckBox">
+
+    </GridLayout>
+
+
+</RelativeLayout>
